Add Finisheh by Saad

+++ app/Http/Controllers/Api/UserController.php
@@ -3,13 +3,13 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Hash;
 use App\Models\User;
-use Tymon\JWTAuth\Facades\JWTAuth;
-use Tymon\JWTAuth\Exceptions\JWTExceptions;
-use Tymon\JWTAuth\Exceptions\UserNotDefinedException;
+use Illuminate\Http\Request;
+// use Illuminate\Support\Facades\Hash;
+use Tymon\JWTAuth\Exceptions\JWTException;
 use Tymon\JWTAuth\Exceptions\TokenInvalidException;
+use Tymon\JWTAuth\Exceptions\UserNotDefinedException;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 
 class UserController extends Controller
@@ -80,8 +80,12 @@ class UserController extends Controller
             return response()->json(['error' => $e->getMessage()]);
         }
             //Select* From Users
+        if($user->role =="admin"){
         $user= User::all();
         return $user;
+            }else{
+                return "You Are Not Admin";
+            }
     }
 
     public function update(Request $request,$id){
@@ -108,6 +112,9 @@ class UserController extends Controller
                 $response['message'] = 'User not found';
                 $response['code'] = 404;
             }
+        }
+        else{
+            return "You Are Not Admin";
         }
         return response()->json($response);
     }
@@ -151,8 +158,14 @@ class UserController extends Controller
         }catch(UserNotDefinedException $e){
             return response()->json(['error' => $e->getMessage()]);
         }
+        if($user->role =="admin"){
+
         $user = User::destroy($id);
         return $user;
+        }
+        else{
+            return "You Are Not Admin";
+        }
     }
